Moved the frequencies property from to the sensor_model class. manipulate the frequencies property such that the band is limited by [MinActiveFreq,MaxActiveFreq]. If [MinActiveFreq,MaxActiveFreq] is larger than the the define frequency band, the band is extended to cover. Added a frequency property to sensor_pair. At the time that the sensor_pair is created, defined its frequency property as the intersection of the source and receiver frequency properties. Changed sensor_pair eigenrays to fathometer_model.

+++ waveq3d/test/reflection_test.cc
@@ -34,8 +34,8 @@ public:
     double latitude ;
     bool surface ;
     bool bottom ;
-    int surf_count ;
-    int bot_count ;
+    size_t surf_count ;
+    size_t bot_count;
 
     /**
      * Initialize counter.
@@ -46,7 +46,7 @@ public:
 
     virtual ~reflection_callback() {}
 
-    bool check_count( int& old_count ) {
+    bool check_count( size_t& old_count ) {
         bottom = surface = false ;
         if( surf_count != _collection->eigenverbs(1).size() ) {
             surf_count = _collection->eigenverbs(1).size() ;
@@ -149,7 +149,7 @@ BOOST_AUTO_TEST_CASE( reflect_flat_test ) {
         eigenverb_collection verb_collection( 0 ) ;
         wave.add_eigenverb_listener( &verb_collection ) ;
         reflection_callback callback( &verb_collection ) ;
-        int old_count = callback.surf_count + callback.bot_count ;
+        size_t old_count = callback.surf_count + callback.bot_count ;
         double max_time_error = 0.0 ;
         double max_lat_error = 0.0 ;
 
@@ -214,7 +214,7 @@ BOOST_AUTO_TEST_CASE( reflect_flat_test ) {
             if( callback.check_count( old_count ) ) {
                 double predict_time = old_count * 7.450560973 ;
                 double current_time = callback.time ;
-                double predict_lat = 45.0 + old_count * 0.1 ;
+                double predict_lat = 45.0 + int(old_count) * 0.1 ;
                 double current_lat = callback.latitude ;
 
                 if ( callback.surface ) cout << "surface" ;
